+++ external/anomalib/callbacks/__init__.py
