+++ external/anomalib/callbacks/inference.py
