+++ external/anomalib/callbacks/model_monitor.py
