+++ external/anomalib/callbacks/progress.py
