+++ external/anomalib/configs/stfpm/template.yaml
@@ -1,5 +1,4 @@
-# Description.
-name: ote anomaly classification padim
+name: ote anomaly classification stfpm
 task_type: ANOMALY_CLASSIFICATION
 task_family: VISION
 instantiation: "CLASS"
@@ -16,11 +15,11 @@ entrypoints:
 
 # Hyperparameters.
 hyper_parameters:
-  base_path: ../../configuration.yaml
+  base_path: ../../anomaly_classification/src/configuration.yaml
   parameter_overrides:
     model:
       name:
-        default_value: padim
+        default_value: stfpm
     dataset:
       train_batch_size:
         default_value: 32
