+++ external/anomalib/tests/test_ote_anomaly_classification.py
@@ -21,7 +21,7 @@ from threading import Thread
 import pytest
 import numpy as np
 
-from anomaly_classification.src.utils.config import get_anomalib_config
+from utils.config import get_anomalib_config
 from tests.helpers.config import get_config
 from tests.helpers.train import OTEAnomalyTrainer
 
@@ -44,7 +44,7 @@ class TestAnomalyClassification:
         """
         train_batch_size = 16
 
-        ote_config = get_config(f"anomaly_classification/src/configs/{template_path}/template.yaml")
+        ote_config = get_config(f"configs/{template_path}/template.yaml")
 
         # change parameter value in OTE config
         ote_config.dataset.train_batch_size = train_batch_size
@@ -57,7 +57,7 @@ class TestAnomalyClassification:
         """
         Training should stop when `cancel_training` is called
         """
-        self._trainer = OTEAnomalyTrainer(model_template_path=f"anomaly_classification/src/configs/{template_path}/template.yaml")
+        self._trainer = OTEAnomalyTrainer(model_template_path=f"configs/{template_path}/template.yaml")
         thread = Thread(target=self._trainer.train)
         thread.start()
         self._trainer.cancel_training()
@@ -68,7 +68,7 @@ class TestAnomalyClassification:
         E2E Train-Export Should Yield Similar Inference Results
         """
         # Train the model
-        self._trainer = OTEAnomalyTrainer(model_template_path=f"anomaly_classification/src/configs/{template_path}/template.yaml")
+        self._trainer = OTEAnomalyTrainer(model_template_path=f"configs/{template_path}/template.yaml")
         self._trainer.train()
         base_results = self._trainer.validate(task=self._trainer.base_task)
