+++ external/anomalib/utils/data.py
